Read multiple filters in one stream.
ADD: We can now read JCSDL of multiple filters. It will create a filter row for each filter in the list of filters.

ADD: The whole editor is now wrapped with an editor element and this initializes automatically on creation of the JCSDLGui instance.
ADD: JCSDLGui.init() can also be used to fully reinitialize the editor.
CHG: Reading the value from JCSDL no longer includes brackets and quotes.
CHG: We should no longer use id attributes on any of the elements.